docs/requirements_docs.txt
matplotlib
numpydoc
pydata-sphinx-theme==0.13.3
scipy
sphinx
sphinx-autobuild
sphinx-design
sphinx-togglebutton